Pimpled MDI item
This should make it easier to extend the mdi item later without ABI changes.

Note: This is an ABI change!
